Refactor Multi._delay as Multi._deadline
Rather than a millisecond delay to sleep for at some point in the
future, '_deadline' is an absolute time according to the monotonic
process clock to sleep until, in millisecond precision.

Give the speed at which the timer is likely to be started after setting
the time it will not make a lot of difference to the timing accuracy.